examples/transitions/pyproject.toml
[tool.poetry]
name = "example"
version = "0.0.0"
description = "example with dependencies"
authors = []

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.24.3"
pyelftools = "^0.29"

[tool.poetry.group.dev.dependencies]
pytest = {version = "^7.2", "python" = ">=3.7"}
